backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
neo4j>=5.15.0
pydantic

langchain
langchain-experimental
langchain-community
langchain-openai
langchain-neo4j>=0.0.9

pypdf
markdown2
python-multipart
requests
stix2
spacy
tiktoken
doc2txt